+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email "bflad417@gmail.com"
 license          "Apache 2.0"
 description      "Installs/Configures jmxtrans"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.4.1"
+version          "0.5.0"
 
 %w{ ark }.each do |cb|
   depends cb